fix: Consistent wp usage in raster RGB extraction (fixes #1349) (#1350)
fix: Consistent wp usage in raster RGB extraction (fixes #1349)

Summary
- Replace real(real64) with real(wp) and remove redundant local
iso_fortran_env imports in raster RGB extraction routines.
- Aligns with repo style and avoids kind mismatches if wp changes.

Changes
- src/backends/raster/fortplot_raster_rendering.f90: use wp in
raster_extract_rgb_data
- src/backends/raster/fortplot_raster.f90: use wp in
raster_extract_rgb_data_context

Verification
Commands run:
- make build
- make test-ci
- make verify-artifacts

Key outputs:
- ALL TESTS PASSED (fpm test)
- CI essential test suite completed successfully
- Artifact verification passed (examples under output/example/fortran/*)
- e.g., output/example/fortran/pcolormesh_demo/pcolormesh_basic.pdf
contains RGB Image XObject
- ylabel stripe checks passed for
output/example/fortran/ylabel_left/*.png

No functional changes observed; this is a consistency cleanup.


Reviewer updates
- Also aligned wp usage in animation extract API and ASCII backend
support.
- Rebuilt, re-ran test-ci and verify-artifacts; all passed locally.

+++ src/animation/fortplot_animation_rendering.f90
@@ -1,5 +1,5 @@
 module fortplot_animation_rendering
-    use iso_fortran_env, only: real64, wp => real64
+    use iso_fortran_env, only: wp => real64
     use fortplot_figure_core, only: figure_t, plot_data_t
     use fortplot_utils, only: initialize_backend
     implicit none
@@ -14,7 +14,7 @@ contains
 
     subroutine extract_frame_rgb_data(fig, rgb_data, status)
         type(figure_t), intent(inout) :: fig
-        real(real64), intent(out) :: rgb_data(:,:,:)
+        real(wp), intent(out) :: rgb_data(:,:,:)
         integer, intent(out) :: status
         
         status = 0
